+++ ios/PostbackBridge.m
@@ -56,6 +56,9 @@ RCT_EXTERN_METHOD(destroy:(RCTPromiseResolveBlock)resolve
 RCT_EXTERN_METHOD(getDeviceInfo:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 
+RCT_EXTERN_METHOD(getWebViewUserAgent:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+
 RCT_EXTERN_METHOD(getAdServicesToken:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)