+++ src/NativePostback.ts
@@ -43,6 +43,9 @@ const unsupportedPlatformModule: NativePostbackModule = {
   async getDeviceInfo() {
     return {};
   },
+  async getWebViewUserAgent() {
+    return null;
+  },
   async getAdServicesToken() {
     return null;
   },